requirements.txt
Flask~=1.1.2
flask_login
redis
mongo
psycopg2
SQLAlchemy~=1.3.17
Flask-WTF
flask_uploads
passlib~=1.7.2
WTForms~=2.3.1
Pillow~=7.0.0
Werkzeug==0.16.0
flask_session
